+++ tests/lib.rs
@@ -172,9 +172,14 @@ testcase!(test_haplotype_singleton, exact);
 testcase!(test_alt_locus_bias_01, exact);
 testcase!(test_alt_locus_bias_02, exact);
 testcase!(test_uzuner_fn_mnv, exact);
+testcase!(test_uzuner_fp_mnv1, exact);
 
 testcase!(test_imprecise_fusion, exact);
 testcase!(test_imprecise_fusion_absent, exact);
+
+testcase!(test_uzuner_clonal_1, exact);
+testcase!(test_uzuner_clonal_2, exact);
+testcase!(test_uzuner_clonal_3, exact);
 
 fn basedir(test: &str) -> String {
     format!("tests/resources/{}", test)